+++ src/app/components/event-log-entry/entry-parsers/openvr-device-turned-off.ts
@@ -1,0 +1,16 @@
+import { EventLogEntryParser } from '../event-log-entry-parser';
+import { EventLogTurnedOffOpenVRDevices, EventLogType } from '../../../models/event-log-entry';
+
+export class EventLogTurnedOffOpenVRDevicesEntryParser extends EventLogEntryParser<EventLogTurnedOffOpenVRDevices> {
+  entryType(): EventLogType {
+    return 'turnedOffOpenVRDevices';
+  }
+
+  override headerInfoTitle(entry: EventLogTurnedOffOpenVRDevices): string {
+    return 'comp.event-log-entry.type.turnedOffOpenVRDevices.title.' + entry.devices;
+  }
+
+  override headerInfoSubTitle(entry: EventLogTurnedOffOpenVRDevices): string {
+    return 'comp.event-log-entry.type.turnedOffOpenVRDevices.reason.' + entry.reason;
+  }
+}