+++ src/app/models/event-log-entry.ts
@@ -1,3 +1,4 @@
+import { LighthouseDevicePowerState } from './lighthouse-device';
 import { SleepModeStatusChangeReason } from './sleep-mode';
 import { UserStatus } from 'vrchat/dist';
 
@@ -7,14 +8,15 @@ export type EventLog = {
 };
 
 export const EVENT_LOG_DEFAULT: EventLog = {
-  version: 1,
+  version: 2,
   logs: [],
 };
 
 export type EventLogEntry =
   | EventLogSleepModeEnabled
   | EventLogSleepModeDisabled
-  | EventLogTurnedOffDevices
+  | EventLogTurnedOffOpenVRDevices
+  | EventLogLighthouseSetPowerState
   | EventLogGpuPowerLimitChanged
   | EventLogBrightnessChanged
   | EventLogAcceptedInviteRequest
@@ -28,7 +30,8 @@ export type EventLogDraft = Omit<EventLogEntry, 'time' | 'id'>;
 export type EventLogType =
   | 'sleepModeEnabled'
   | 'sleepModeDisabled'
-  | 'turnedOffDevices'
+  | 'turnedOffOpenVRDevices'
+  | 'lighthouseSetPowerState'
   | 'gpuPowerLimitChanged'
   | 'brightnessChanged'
   | 'acceptedInviteRequest'
@@ -54,10 +57,17 @@ export interface EventLogSleepModeDisabled extends EventLogBase {
   reason: SleepModeStatusChangeReason;
 }
 
-export interface EventLogTurnedOffDevices extends EventLogBase {
-  type: 'turnedOffDevices';
+export interface EventLogTurnedOffOpenVRDevices extends EventLogBase {
+  type: 'turnedOffOpenVRDevices';
   reason: 'MANUAL' | 'OSC_CONTROL' | 'SLEEP_MODE_ENABLED' | 'CHARGING';
   devices: 'CONTROLLER' | 'CONTROLLERS' | 'TRACKER' | 'TRACKERS' | 'ALL' | 'VARIOUS';
+}
+
+export interface EventLogLighthouseSetPowerState extends EventLogBase {
+  type: 'lighthouseSetPowerState';
+  reason: 'MANUAL' | 'OYASUMI_START' | 'STEAMVR_START' | 'STEAMVR_STOP';
+  devices: 'ALL' | 'SINGLE';
+  state: LighthouseDevicePowerState;
 }
 
 export interface EventLogGpuPowerLimitChanged extends EventLogBase {